requirements.txt
streamlit
streamlit-folium
folium
requests
pandas
pyTelegramBotAPI
google-generativeai
Pillow